Fixed missing GST slabs and enforced whole-number GST rates across system

+++ models/gst.js
@@ -1,3 +1,3 @@
 module.exports = (sequelize, DataTypes) => {
-    return sequelize.define('GST', { rate: { type: DataTypes.DECIMAL(5, 2), unique: true } });
+    return sequelize.define('GST', { rate: { type: DataTypes.INTEGER, unique: true } });
 };
